+++ src/TreeView/index.ts
@@ -1,5 +1,5 @@
 /* ======================================================================== *
- * Copyright 2024 HCL America Inc.                                          *
+ * Copyright 2026 HCL America Inc.                                          *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
  * You may obtain a copy of the License at                                  *
@@ -12,15 +12,8 @@
  * See the License for the specific language governing permissions and      *
  * limitations under the License.                                           *
  * ======================================================================== */
-import React from 'react';
-import MuiTreeItem, { TreeItemProps } from '@mui/lab/TreeItem';
 
-const TreeItem = ({ ...props }: TreeItemProps) => {
-  return <MuiTreeItem {...props} />;
-};
+import TreeView from './TreeView';
 
-TreeItem.defaultProps = {
-};
-
-export * from '@mui/lab/TreeItem';
-export default TreeItem;
+export default TreeView;
+export * from './TreeView';